CveSearch: refactor SearchLevel generation
