+++ backend/src/src-cvesearch/src/main/java/com/bosch/osmi/sw360/cvesearch/datasource/matcher/Match.java
@@ -29,4 +29,9 @@ public class Match {
     public Match concat(Match otherMatch) {
         return new Match(this.needle + ":" + otherMatch.getNeedle(), this.distance + otherMatch.getDistance());
     }
+
+    @Override
+    public String toString() {
+        return "[" + this.needle + ":" + this.distance + "]";
+    }
 }
